Adding Customer class with basic CRUD operations using DAO

+++ src/main/java/com/projectIO/touritsitcEquipmentRentalShop/dao/ItemCategoryDao.java
@@ -4,8 +4,8 @@ import com.projectIO.touritsitcEquipmentRentalShop.model.ItemCategory;
 
 public interface ItemCategoryDao {
     void save(ItemCategory itemCategory);
-    ItemCategory read(Long id);
+    ItemCategory read(Integer id);
     void update(ItemCategory itemCategory);
-    void delete(Long id);
+    void delete(Integer id);
     void cleanUp();
 }
