+++ src/main/java/com/projectIO/touritsitcEquipmentRentalShop/dao/ItemDao.java
@@ -4,8 +4,8 @@ import com.projectIO.touritsitcEquipmentRentalShop.model.Item;
 
 public interface ItemDao {
     void save(Item item);
-    Item read(Long id);
+    Item read(Integer id);
     void update(Item item);
-    void delete(Long id);
+    void delete(Integer id);
     void cleanUp();
 }
